🔨 refactor: ga 로드 코드 추출
