+++ src/app/App.tsx
@@ -1,33 +1,13 @@
-import { lazy, Suspense, useEffect } from 'react';
+import { lazy, Suspense } from 'react';
 import { Route, Routes } from 'react-router-dom';
 
 import Home from '@pages/home';
+import { useGoogleAnalytics } from '@shared/lib/analytics/useGoogleAnalytics';
 
 const PrivacyPolicy = lazy(() => import('@pages/privacy-policy'));
 
-const GA_ID = 'G-JMQZ9VKYZ6';
-
-declare global {
-  interface Window {
-    dataLayer: unknown[];
-    gtag: (...args: unknown[]) => void;
-  }
-}
-
 function App() {
-  useEffect(() => {
-    const script = document.createElement('script');
-    script.src = `https://www.googletagmanager.com/gtag/js?id=${GA_ID}`;
-    script.async = true;
-    document.head.appendChild(script);
-
-    window.dataLayer = window.dataLayer || [];
-    window.gtag = function (...args: unknown[]) {
-      window.dataLayer.push(args);
-    };
-    window.gtag('js', new Date());
-    window.gtag('config', GA_ID);
-  }, []);
+  useGoogleAnalytics();
 
   return (
     <Routes>
