+++ src/shared/lib/analytics/useGoogleAnalytics.ts
@@ -1,0 +1,26 @@
+import { useEffect } from 'react';
+
+declare global {
+  interface Window {
+    dataLayer: unknown[];
+    gtag: (...args: unknown[]) => void;
+  }
+}
+
+const GA_ID = 'G-JMQZ9VKYZ6';
+
+export function useGoogleAnalytics() {
+  useEffect(() => {
+    const script = document.createElement('script');
+    script.src = `https://www.googletagmanager.com/gtag/js?id=${GA_ID}`;
+    script.async = true;
+    document.head.appendChild(script);
+
+    window.dataLayer = window.dataLayer || [];
+    window.gtag = function (...args: unknown[]) {
+      window.dataLayer.push(args);
+    };
+    window.gtag('js', new Date());
+    window.gtag('config', GA_ID);
+  }, []);
+}
